v2.5
Added the /BeamOut/Source command to simulate stationary "beam"
particles so that level scheme files can be used to simulate gamma-ray
cascades in stationary sources. The level scheme file is based on
measured beta- and gamma-decay branching ratios (in the level scheme
file) which combine to yield innacurate gamma-ray intensities in
simulated spectra. Hence, this approach should not be used to
determine gamma-ray efficiencies.

The existing source simulation facilities remain in place for use in
efficiency calibrations.

Removed the tracking method option.