v2.6
Full relativistic 2-body (2 in, 2 out) reaction kinematics are now
used. Target (A,Z) must be specified with the /BeamOut/TargetA and
/BeamOut/TargetZ commands.

Added an option (/Mode2/crystalXforms) to use internal transformations
from the world frame to the crystal frames for Mode2 output. There are
small discrepancies between the geometry files and the design
positions of the cryrstals such that the transformations specified in
the crmat.LINUX file result in crystal-frame hit patterns that are
offset by as much as a few mm.

Added glassy carbon "glassyC" to the available target materials.