v2.7
Updated the EventInformation class to include the velocity of the
projectile at each gamma emmission and a full-energy flag, set to 1
for events in which a single gamma ray is emitted and its full energy
is deposited in a single crystal, 0 for events in which a single gamma
ray is emitted and only part of its energy is deposited in any one
crystal, and -1 otherwise.  Added the full-energy flag and the
projectile beta values to the output stream.

WARNING: the data structure in type 11 packets in Mode 2
output has changed. Sorting codes that read these packets will need to
be updated. The relevant data structure (g4sim_emitted_gamma) is
defined in ./include/GEB.hh.

Updated the ASCII output format to write the same information as
Mode 2 output (S800 ATA, BTA, DTA, and YTA, decomposed gamma-ray
events, and descriptions of emitted gamma rays). The new format is
described in the README.md file.

Changed MAX_SIM_GAMMAS to 10 to reduce output file size. Compile
with HIGHMULT=1 to set this to 40 for high multiplicity simulations.

Added "Sn" target material.

Some minor cleanup