v3.0
New geometry files correcting slight discrepancies with actual
positions. The new cluster and wall files center the template cluster
on the z axis. The euler files position and orient the modules based
on the crmat transformations from crystal to world coordinates.  (This
is more straightforward than euler angles relative to a template
cluster placed in slot 1.)