Added the optional /Experiment/Source/ThetaFile macro-file command which used a user-provided theta distribution to set the direction of emitted particles (works with the "simple" and "neutron" source types).
Added optional /Experiment/Source/setDX, /Experiment/Source/setDY,
/Experiment/Source/setSigmaX, and /Experiment/Source/setSigmaY
macro-file commands for simulating a rectangular flat or Gaussian
source profile. The flat/Gaussian options can be mixed in x and y.

The (existing, optional) /Experiment/Source/setWhiteLowE and
/Experiment/Source/setWhiteLowE macro-file commands can now be used
with the "simple" and "neutron" source types in addition to the
"white" and "bgwhite" types.