Replaced the algorithm for setting the z position at which the reaction takes place in the target.
The existing method would sometimes find that the target had zero thickness
and emit gammas from the center of the target. The existing method relied on
the DistanceToIn() method of the G4UnionSolid class to find the point at
which the beam enters the target. This sometimes yielded points of entry not
in the solid (roundoff error?). Then, the DistanceToOut() method
would fail silently when passed a point outside of the solid.

A more robust method is now used, developed with Jake Kosa and Alyssa
Himmelreich, which uses the DistanceToOut() method from the point of
intersection of the beam trajectory with the central plane of the target.